fix race conditions during downloads (#114)

+++ chai_lab/utils/paths.py
@@ -4,10 +4,12 @@
 
 import dataclasses
 import os
+import random
 from pathlib import Path
 from typing import Final
 
 import requests
+from filelock import FileLock
 
 # use this path object to specify location
 # of anything within repository
@@ -23,19 +25,24 @@ downloads_path = Path(os.environ.get("CHAI_DOWNLOADS_DIR", downloads_path))
 assert repo_root.exists()
 
 
-def download(http_url: str, path: Path):
+def download_if_not_exists(http_url: str, path: Path):
     print(f"downloading {http_url}")
-    tmp_path = path.with_suffix(".download_tmp")
+    if path.exists():
+        return
 
-    with requests.get(http_url, stream=True) as response:
-        response.raise_for_status()  # Check if the request was successful
-        # Open a local file with the specified name
-        path.parent.mkdir(exist_ok=True, parents=True)
-        with tmp_path.open("wb") as file:
-            # Download the file in chunks
-            for chunk in response.iter_content(chunk_size=8192):
-                if chunk:  # Filter out keep-alive new chunks
-                    file.write(chunk)
+    with FileLock(path.with_suffix(".download_lock")):
+        if path.exists():
+            return  # if-lock-if sandwich to download only once
+        tmp_path = path.with_suffix(f".download_tmp_{random.randint(10 ** 5, 10**6)}")
+        with requests.get(http_url, stream=True) as response:
+            response.raise_for_status()  # Check if the request was successful
+            # Open a local file with the specified name
+            path.parent.mkdir(exist_ok=True, parents=True)
+            with tmp_path.open("wb") as file:
+                # Download the file in chunks
+                for chunk in response.iter_content(chunk_size=8192):
+                    if chunk:  # Filter out keep-alive new chunks
+                        file.write(chunk)
     tmp_path.rename(path)
     assert path.exists()
 
@@ -47,9 +54,7 @@ class Downloadable:
 
     def get_path(self) -> Path:
         # downloads artifact if necessary
-        if not self.path.exists():
-            download(self.url, path=self.path)
-
+        download_if_not_exists(self.url, path=self.path)
         return self.path
 
 
@@ -68,6 +73,6 @@ def chai1_component(comp_key: str) -> Path:
     url = f"https://chaiassets.com/chai1-inference-depencencies/models/{comp_key}"
     result = downloads_path.joinpath("models", comp_key)
     if not result.exists():
-        download(url, result)
+        download_if_not_exists(url, result)
 
     return result
